Removed all matter.js code. Refactored Game2 to Game.
