+++ frontend/src/components/Game/Game.js
@@ -1,116 +1,124 @@
 import React from "react";
 import io from "socket.io-client";
-import {Bodies, Body, Composite, Engine, Events, Render, World} from "matter-js";
-import {BACKEND_URL} from "config";
-import {canvasWidth, canvasHeight, boundWidth, paddleBoundSpacing, paddleOffset} from "./utils/dimensions";
-import colors from "./utils/colors";
+import {World} from "p2";
+import P5 from "p5";
 import Controls from "./controls";
-import {Player, PaddleBoundPair} from "./bodies";
-import "./Game.css";
+import {BACKEND_URL} from "config";
+import {canvasWidth, canvasHeight} from "./utils/dimensions";
+import colors from "./utils/colors";
+import {worldSetup, addBall, addPlayer} from "./worldSetup";
 
 export default class Game extends React.Component {
 
     constructor() {
         super();
-        this.engine = Engine.create();
 
         this.socket = io.connect(BACKEND_URL, {path: "/game-socket"});
-
-        // create two players
-        this.player = new Player(paddleOffset + paddleBoundSpacing/2, canvasHeight/2);
-        this.opponent = new Player(canvasWidth - paddleOffset, canvasHeight/2);
-        
-        // make the players face each other
-        Composite.rotate(this.player, Math.PI/2, {x: paddleOffset, y: canvasHeight/2});
-        Composite.rotate(this.opponent, -Math.PI/2, {x: canvasWidth - paddleOffset, y: canvasHeight/2});
-
         this.controls = new Controls(this.socket); //TODO: Remove dependency on socket from controls
 
-        this.createWorld(); //TODO: move create world, players, and rotation to a separate world initialization file
+        this.world = new World({
+            gravity: [0, 0]
+        });
+
+        this.world.defaultContactMaterial.restitution = 1;
+        this.world.defaultContactMaterial.contactSkinSize = 0;
     }
 
     componentDidMount() {
-        // create the renderer
-        this.render = Render.create({
-            element: document.getElementById("game-render"), //attaches game canvas to the div#game-render HTML element
-            engine: this.engine,
-            options: {
-                background: colors.background,
-                height: canvasHeight,
-                width: canvasWidth,
-                wireframes: false
-            }
-        });
-        // autofocus on the renderer
+
+        // Create boundaries
+        worldSetup(this.world);
+        
+        // Ball will eventually be created once two players join.
+        addBall(this.world);
+
+        // Create players
+        const player1 = addPlayer(this.world, 0);
+        const player2 = addPlayer(this.world, 1);
+        
+        // Create sketch
+        const sketch = (p) => { // TODO: Move sketch to its own file; import it here.
+
+            p.setup = () => {
+                p.createCanvas(canvasWidth, canvasHeight);
+                p.push();
+                p.noStroke();
+                p.angleMode(p.RADIANS);
+                p.rectMode(p.CENTER);
+            };
+
+            p.draw = () => {
+                p.background(colors.background);
+                p.fill(colors.defaultColor);
+                for (const body of this.world.bodies) {
+                    body.render(p);
+                }
+                for (const spring of this.world.springs) {
+                    spring.render(p);
+                }
+            };
+        };
+
+        // Create and autofocus on renderer
+        this.p5 = new P5(sketch, document.getElementById("game-render"));
         document.getElementById("game-render").focus();
 
-        this.run();
-    }
+        this.runEngine(this.world);
 
-    createWorld() {
-        this.engine.world.gravity.y = 0;
-
-        // create two boxes and a ground
-        let ball = Bodies.circle(60, canvasHeight/2, 10, {
-            friction: 0,
-            frictionAir: 0
-        });
-        let boxA = Bodies.rectangle(400, 200, 80, 80);
-        let boxB = Bodies.rectangle(350, 100, 50, 50, {
-            render: {
-                fillStyle: "white",
-                strokeStyle: "blue",
-                lineWidth: 3
-            },
-        });
-        boxA.restitution = 0.9;
-        boxB.restitution = 0.9;
-
-        // create thick borders, but ensure only 10px of thickness are visible
-        const ceiling = Bodies.rectangle(canvasWidth/2, 0, canvasWidth + boundWidth, boundWidth, {isStatic: true});
-        const ground = Bodies.rectangle(canvasWidth/2, canvasHeight, canvasWidth + boundWidth, boundWidth, {isStatic: true});
-
-        const player1Bounds = new PaddleBoundPair(paddleOffset + paddleBoundSpacing/2);
-        const player2Bounds = new PaddleBoundPair(canvasWidth - paddleOffset - paddleBoundSpacing/2);
-
-        Events.on(this.engine, "beforeUpdate", (event) => {
-            // react to key commands and apply force as needed
+        // Update the character controller after each physics tick.
+        this.world.on('postStep', () => {
+            const paddle = player1.paddle;
+            const bumper = player1.bumper;
             if (this.controls.UP) {
-                let paddle = this.player.bodies[0];
-                Body.setVelocity(paddle, {x: 0, y: -4});
+                const paddleVelocity = [-30, 0];
+                paddle.vectorToWorldFrame(paddle.velocity, paddleVelocity);
+                let currentLocal = [0, 0];
+                bumper.vectorToLocalFrame(currentLocal, bumper.velocity);
+                let desiredLocal = [-30, currentLocal[1]];
+                bumper.vectorToWorldFrame(bumper.velocity, desiredLocal);
             } else if (this.controls.DOWN) {
-                let paddle = this.player.bodies[0];
-                Body.setVelocity(paddle, {x: 0, y: 4});
+                const paddleVelocity = [30, 0];
+                paddle.vectorToWorldFrame(paddle.velocity, paddleVelocity);
+                let currentLocal = [0, 0];
+                bumper.vectorToLocalFrame(currentLocal, bumper.velocity);
+                let desiredLocal = [30, currentLocal[1]];
+                bumper.vectorToWorldFrame(bumper.velocity, desiredLocal);
             } else {
-                let paddle = this.player.bodies[0];
-                Body.setVelocity(paddle, {x: 0, y: 0});
+                const paddleVelocity = [0, 0];
+                paddle.vectorToWorldFrame(paddle.velocity, paddleVelocity);
+                let currentLocal = [0, 0];
+                bumper.vectorToLocalFrame(currentLocal, bumper.velocity);
+                let desiredLocal = [0, currentLocal[1]];
+                bumper.vectorToWorldFrame(bumper.velocity, desiredLocal);
             }
             if (this.controls.BOOST) {
-                let paddle = this.player.bodies[0];
-                let bumper = this.player.bodies[1];
-                let force = (0.001  * paddle.mass);
-                let forceAngle = paddle.angle - Math.PI/2;
-                let xForce = Math.cos(forceAngle) * force;
-                let yForce = Math.sin(forceAngle) * force; 
-                Body.applyForce(bumper, bumper.position, {x: xForce, y: yForce})
+                const force = 128 * bumper.mass;
+                bumper.applyForceLocal([0, force]);
             }
         });
 
-        // add all of the bodies to the world
-        World.add(this.engine.world, [this.player, this.opponent, player1Bounds, player2Bounds, ball, boxA, boxB, ceiling, ground]);
+        console.log(this.world);
     }
 
-    run() {
-        // run the engine
-        Engine.run(this.engine);
-
-        // run the renderer
-        Render.run(this.render);
+    runEngine(world) {
+        let maxSubSteps = 10;
+        let fixedTimeStep = 1/60;
+        let lastTimeSeconds;
+        const animate = (t) => {
+            requestAnimationFrame(animate);
+            const timeSeconds = t/1000;
+            lastTimeSeconds = lastTimeSeconds || timeSeconds;
+            const timeSinceLastCall = timeSeconds - lastTimeSeconds;
+            world.step(fixedTimeStep, timeSinceLastCall, maxSubSteps);
+        }
+        requestAnimationFrame(animate);
     }
+
+    onMouseUp = (mouseEvent) => addBall(this.world, this.p5.mouseX, canvasHeight - this.p5.mouseY);
 
     render() {
         return (
-            <div id="game-render" tabIndex="0" onKeyDown={this.controls.handleKeyDown} onKeyUp={this.controls.handleKeyUp}>
+            <div id="game-render" tabIndex="0" onKeyDown={this.controls.handleKeyDown} onKeyUp={this.controls.handleKeyUp} onMouseUp={this.onMouseUp}>
                 {
                     //game will be rendered here when the component mounts.
                 }
