+++ frontend/src/components/Game/bodies/boundary.js
