+++ frontend/src/components/Game/bodies/bumpers/curvedBumper.js
