+++ frontend/src/components/Game/bodies/bumpers/index.js
