+++ frontend/src/components/Game/bodies/bumpers/rectangularBumper.js
