+++ frontend/src/components/Game/bodies/bumpers/triangularBumper.js
