+++ frontend/src/components/Game/bodies/player.js
@@ -1,63 +1,146 @@
-import {Bodies, Composite, Constraint} from "matter-js";
-import {paddleBoundSpacing, canvasHeight} from "../utils/dimensions";
-import filters from "../utils/filters";
+import {Body, Box, LinearSpring, DistanceConstraint} from "p2";
+import {canvasHeight, paddleHeight} from "../utils/dimensions"
+import {drawRectangle, drawLineSpring} from "../render";
+import {CurvedBumper} from "./bumpers";
+import colors from "../utils/colors";
+import {groups, masks} from "../utils/collisions";
+import {rotateAroundAnchor} from "../utils/math";
 
 /*
- * By default, a player is horizontal, with the bumper facing upwards
+ * A player is composed of a rectangular paddle and a bumper connected by springs.
+ * The default orientation of a player is bumper facing upwards.
  */
-const player = (x, y) => {
+export default class Player {
+    constructor(x, y, angle) {
+        this.angle = angle
+        // Dimensions
+        this.springLength = paddleHeight;
+        
+        // Bodies
+        this.paddle = new Paddle({
+            position: [x, y],
+            angle: angle
+        });
 
-    // size constants
-    const player = Composite.create({label: 'player'});
-    const width = canvasHeight/5;
-    const paddleHeight = paddleBoundSpacing;
-    const bumperHeight = paddleHeight*2;
-    const springLength = paddleHeight + bumperHeight/2;
+        this.bumper = new CurvedBumper({
+            position: rotateAroundAnchor(angle, x, y + this.springLength, ...this.paddle.position),
+            angle: angle,
+            mass: 100
+        });
 
-    const paddle = Bodies.rectangle(x, y, width, paddleHeight, {
-        collisionFilter: {
-            category: filters.paddles,
-            mask: filters.defaultFilter | filters.invisible
-        },
-        density: 0.1,
-        label: "paddle"
-    });
+        // Springs
+        const spring1 = new PlayerSpring(this.paddle, this.bumper, {
+            localAnchorA: [35, this.paddle.shapes[0].centerOfMass[1]],
+            localAnchorB: [35, this.bumper.shapes[0].centerOfMass[1]],
+            restLength: this.springLength
+        });
+        const spring2 = new PlayerSpring(this.paddle, this.bumper, {
+            localAnchorA: [-35, this.paddle.shapes[0].centerOfMass[1]],
+            localAnchorB: [-35, this.bumper.shapes[0].centerOfMass[1]],
+            restLength: this.springLength
+        });
+        this.springs = [spring1, spring2];
 
-    // Note: bumper is rotated 90 degrees so height is what appears as width in-game
-    const bumper = Bodies.trapezoid(x, y-springLength, width, bumperHeight, 0.1, {
-        chamfer: {
-            radius: [0, 15, 15, 0]
-        },
-        label: "bumper",
-        restitution: 1
-    });
+        // Permanent Constraints
+        const distanceConstraint = new PaddleBumperDistanceConstraint(this.paddle, this.bumper, {
+            upperLimitEnabled: true,
+            lowerLimitEnabled: true,
+            upperLimit: this.springLength*2,
+            lowerLimit: this.springLength
+        });
+        this.constraints = [distanceConstraint];
+    }
 
-    const springA1 = createSpring(paddle, bumper, 35, 30, springLength);
-    const springA2 = createSpring(paddle, bumper, 35, 40, springLength);
-    const springB1 = createSpring(paddle, bumper, -35, -40, springLength); 
-    const springB2 = createSpring(paddle, bumper, -35, -30, springLength);
-
-    Composite.addBody(player, paddle);
-    Composite.addBody(player, bumper);
-    Composite.addConstraint(player, springA1);
-    Composite.addConstraint(player, springA2);
-    Composite.addConstraint(player, springB1);
-    Composite.addConstraint(player, springB2);
-
-    return player;
+    addToWorld(world) {
+        // add bodies
+        world.addBody(this.paddle);
+        world.addBody(this.bumper);
+        
+        // add springs
+        for (const spring of this.springs) {
+            world.addSpring(spring);
+        }
+        for (const constraint of this.constraints) {
+            world.addConstraint(constraint);
+        }
+        
+    }
 }
 
-const createSpring = (bodyA, bodyB, xA, xB, length) => Constraint.create({
-    bodyA: bodyA,
-    bodyB: bodyB,
-    length: length,
-    pointA: {x: xA, y: 0},
-    pointB: {x: xB, y: 0},
-    render: {
-        strokeStyle: "black",
-        type: "line"
-    },
-    stiffness: 0.05
-});
+/*
+ * A paddle is the "spine" of the player. The bumper is attached to it via springs.
+ */
+class Paddle extends Body {
+    constructor(props) {
+        // set defaults if not specified in props
+        props = Object.assign({
+            angularDamping: 0,
+            damping: 0,
+            fixedRotation: true,
+            mass: 1000
+        }, props);
 
-export default player;
+        // call parent constructor
+        super(props);
+
+        // add paddle
+        const paddle = new Box({
+            collisionGroup: groups.paddles,
+            collisionMask: masks.paddles,
+            width: canvasHeight/5,
+            height: paddleHeight
+        });
+        this.addShape(paddle);
+    }
+
+    render = (p) => drawRectangle(p, this, colors.paddle1);
+}
+
+/*
+ * Springs with rendering
+ */
+class PlayerSpring extends LinearSpring {
+    constructor(paddle, bumper, props) {
+        // set defaults if not specified in props
+        props = Object.assign({
+            stiffness: 3*bumper.mass,
+            restLength: 10,
+            damping : 0.5,
+            localAnchorA: [-35, paddle.shapes[0].centerOfMass[1]],
+            localAnchorB: [-35, bumper.shapes[0].centerOfMass[1]]
+        }, props);
+
+        // call parent constructor
+        super(paddle, bumper, props);
+    }
+
+    render = (p) => drawLineSpring(p, this, colors.defaultColor);
+}
+
+/*
+ * Limits to how far the springs may stretch
+ */
+class PaddleBumperDistanceConstraint extends DistanceConstraint {
+    constructor(paddle, bumper, props) {
+        // set defaults if not specified in props
+        props = Object.assign({
+            upperLimitEnabled: true,
+            lowerLimitEnabled: true,
+            upperLimit: 100,
+            lowerlimit: 0,
+            localAnchorA: [0, 0],
+            localAnchorB: [0, 0]
+        }, props);
+
+        // call parent constructor
+        super(paddle, bumper, props);
+
+        // Fix upper and lowerlimit props not being applied
+        // Opened pull request: https://github.com/schteppe/p2.js/pull/341
+        this.upperLimitEnabled = props.upperLimitEnabled;
+        this.lowerLimitEnabled = props.lowerLimitEnabled;
+        this.lowerLimit = props.lowerLimit;
+        this.upperLimit = props.upperLimit;
+    }
+}
+
