+++ frontend/src/components/Game/render/index.js
