+++ frontend/src/components/Game/render/springs.js
