+++ frontend/src/components/Game/utils/dimensions.js
@@ -1,13 +1,18 @@
+// canvas
 const canvasHeight = 600;
 const canvasWidth = 900;
-const paddleBoundSpacing = 10;
-const paddleOffset = 10;
+// world
 const boundWidth = 30;
+const paddleOffset = 10;
+// player
+const paddleWidth = canvasHeight/5;
+const paddleHeight = 10;
 
 export {
     canvasHeight,
     canvasWidth,
-    paddleBoundSpacing,
     paddleOffset,
-    boundWidth
+    boundWidth,
+    paddleWidth,
+    paddleHeight
 };
