--- frontend/src/components/Game/utils/filters.js
@@ -1,9 +1,0 @@
-const defaultFilter = 0x0001;
-const invisible = 0x0002;
-const paddles = 0x0004;
-
-export default {
-    defaultFilter,
-    invisible,
-    paddles
-};
