+++ frontend/src/components/Game/utils/math.js
