--- frontend/src/components/Game2/Game.scss
@@ -1,10 +1,0 @@
-#game-render {
-    display: flex;
-
-    canvas {
-        // Modify the CSS of the game canvas
-        margin: 0;
-        padding: 0;
-    }
-
-}
