--- frontend/src/components/Game2/bodies/index.js
@@ -1,9 +1,0 @@
-import Ball from "./ball";
-import Boundary from "./boundary";
-import Player from "./player";
-
-export {
-    Ball,
-    Boundary,
-    Player
-};
