--- frontend/src/components/Game2/index.js
@@ -1,3 +1,0 @@
-import Game from "./Game";
-
-export default Game;
