--- frontend/src/components/Game2/utils/colors.js
@@ -1,21 +1,0 @@
-const defaultColor = "#364040"
-const background = "#ffffff"
-const border = "#2e2b44"
-const paddle1 = "#0496ff"
-const bumper1 = "#d81159"
-const paddle2 = "#008000"
-const bumper2 = "#ffbc42"
-const ball = "#006ba6"
-const ball2 = "#8f2d56"
-
-export default {
-    defaultColor,
-    background,
-    ball,
-    ball2,
-    border,
-    paddle1,
-    bumper1,
-    paddle2,
-    bumper2
-};
