--- frontend/src/components/Game2/utils/dimensions.js
@@ -1,18 +1,0 @@
-// canvas
-const canvasHeight = 600;
-const canvasWidth = 900;
-// world
-const boundWidth = 30;
-const paddleOffset = 10;
-// player
-const paddleWidth = canvasHeight/5;
-const paddleHeight = 10;
-
-export {
-    canvasHeight,
-    canvasWidth,
-    paddleOffset,
-    boundWidth,
-    paddleWidth,
-    paddleHeight
-};
