+++ frontend/src/components/index.js
@@ -1,9 +1,7 @@
 import Game from "./Game";
-import Game2 from "./Game2";
 import Header from "./Header";
 
 export {
     Game,
-    Game2,
     Header
 };
